refactor: spin up a separate Node process for each preview server (#2126)
This enables us to restart Vite completely fresh when a Node dependency
(e.g. a Vite plugin) is updated.

+++ daemon/patch-build.ts
@@ -1,14 +1,26 @@
-// For some reason, index.mjs ends up with an incorrect import
+// For some reason, we end up with an incorrect import
 // that looks like `import ... from 'process/'`.
 //
 // This script removes the unwanted slash.
 
 import fs from "fs";
+import path from "path";
 
-const indexFilePath = "dist/index.mjs";
-const indexContent = fs.readFileSync(indexFilePath, "utf8");
-fs.writeFileSync(
-  indexFilePath,
-  indexContent.replace(`'process/'`, `'process'`),
-  "utf8"
-);
+visit("./dist");
+
+function visit(dirPath: string) {
+  for (const f of fs.readdirSync(dirPath)) {
+    const filePath = path.join(dirPath, f);
+    const stat = fs.statSync(filePath);
+    if (stat.isFile()) {
+      const content = fs.readFileSync(filePath, "utf8");
+      fs.writeFileSync(
+        filePath,
+        content.replace(`'process/'`, `'process'`),
+        "utf8"
+      );
+    } else if (stat.isDirectory()) {
+      visit(filePath);
+    }
+  }
+}
